Final_Project/Demo/backend/application/requirements.txt
aioredis==2.0.1
aiosmtplib==1.1.6
alembic==1.7.7
anyio==3.5.0
appdirs==1.4.3
apturl==0.5.2
asgiref==3.5.1
async-timeout==4.0.2
attrs==19.3.0
bcrypt==3.1.7
blinker==1.4
boto3==1.23.3
botocore==1.26.3
Brlapi==0.7.0
cached-property==1.5.1
certifi==2019.11.28
chardet==3.0.4
charset-normalizer==2.0.12
click==8.0.3
colorama==0.4.3
command-not-found==0.3
cryptography==2.8
cupshelpers==1.0
dbus-python==1.2.16
defer==1.0.6
Deprecated==1.2.13
distlib==0.3.4
distro==1.4.0
distro-info===0.23ubuntu1
dnspython==2.2.1
docker==4.1.0
dockerpty==0.4.1
docopt==0.6.2
duplicity==0.8.12.0
ecdsa==0.17.0
email-validator==1.2.1
entrypoints==0.3
fakeredis==1.7.5
fastapi==0.75.2
fastapi-jwt-auth==0.5.0
fastapi-mail==1.0.8
fasteners==0.14.1
filelock==3.4.2
Flask==2.0.2
future==0.18.2
greenlet==1.1.2
h11==0.12.0
httpcore==0.14.7
httplib2==0.14.0
httpx==0.22.0
idna==2.8
importlib-metadata==1.5.0
importlib-resources==5.7.1
iniconfig==1.1.1
itsdangerous==2.0.1
Jinja2==3.0.2
jmespath==1.0.0
jose==1.0.0
jsonschema==3.2.0
keyring==18.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lockfile==0.12.2
louis==3.12.0
macaroonbakery==1.3.1
Mako==1.1.0
MarkupSafe==2.0.1
monotonic==1.5
more-itertools==4.2.0
netifaces==0.10.4
oauthlib==3.1.0
olefile==0.46
packaging==21.3
paramiko==2.6.0
passlib==1.7.4
pexpect==4.6.0
Pillow==7.0.0
platformdirs==2.5.0
pluggy==1.0.0
protobuf==3.6.1
psycopg2-binary==2.9.3
py==1.11.0
pyasn1==0.4.8
pycairo==1.16.2
pycups==1.9.73
pydantic==1.9.0
PyGObject==3.36.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyparsing==3.0.9
pyRFC3339==1.1
pyrsistent==0.15.5
pytest==7.1.2
python-apt==2.0.0+ubuntu0.20.4.7
python-dateutil==2.7.3
python-debian===0.1.36ubuntu1
python-dotenv==0.20.0
python-jose==3.3.0
python-multipart==0.0.5
pytz==2019.3
pyxdg==0.26
PyYAML==5.3.1
redis==4.3.1
reportlab==3.5.34
requests==2.22.0
requests-unixsocket==0.2.0
rfc3986==1.5.0
rsa==4.8
s3transfer==0.5.2
SecretStorage==2.3.1
simplejson==3.16.0
six==1.14.0
sniffio==1.2.0
sortedcontainers==2.4.0
SQLAlchemy==1.4.36
starlette==0.17.1
systemd-python==234
texttable==1.6.2
tomli==2.0.1
typing-extensions==4.2.0
ubuntu-advantage-tools==27.6
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.25.8
usb-creator==0.3.7
uvicorn==0.17.6
virtualenv==20.13.1
wadllib==1.3.3
websocket-client==0.53.0
Werkzeug==2.0.2
wrapt==1.14.1
xkit==0.0.0
zipp==3.8.0
